+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 ## Shibaraku master (unreleased)
 [full changelog](https://github.com/onk/shibaraku/compare/v0.0.2...master)
 
+*   カラム名が `start_at` / `end_at` 決め打ちだったので、`started_at` 等をオプションで渡せるようにした
+
+    ```ruby
+    class Event
+      shibaraku start_at: :started_at, end_at: :ended_at
+    end
+    ```
+
 
 ## Shibaraku v0.0.2 (2014-12-05)